Rename Openswitch to LibreSwitch
