+++ lib/ansible/utils/module_docs_fragments/libreswitch.py
@@ -88,7 +88,7 @@ options:
     choices: ['yes', 'no']
   provider:
     description:
-      - Convenience method that allows all I(openswitch) arguments to be passed as
+      - Convenience method that allows all I(libreswitch) arguments to be passed as
         a dict object.  All constraints (required, choices, etc) must be
         met either by individual arguments or values in this dict.
     required: false
